NewProject/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
openpyxl
pandas
fuzzywuzzy
python-Levenshtein
xlrd
fhir.resources
requests
